Moce loss in bf16

+++ train.py
@@ -130,7 +130,7 @@ def main(job_config: JobConfig):
     # loss function to be shared by Pipeline Parallel and SPMD training
     def loss_fn(pred, labels):
         return torch.nn.functional.cross_entropy(
-            pred.flatten(0, 1).float(), labels.flatten(0, 1)
+            pred.flatten(0, 1), labels.flatten(0, 1)
         )
 
     def train_step(input_ids, labels):
